# openapi.yaml
openapi: 3.0.1
info:
  title: Orçamento Familiar API
  version: 1.0.0
tags:
  - name: Titulos
    description: Lançamentos de Despesas e Receitas mensal
paths:
  '/titulos/lancamentos/{ano}/{mes}':
    get:
      tags:
        - Titulos
      summary: Serviço para realizar lançamentos de despesas do mes
      operationId: lancamentosTitulo
      parameters:
        - name: ano
          in: path
          required: true
          schema:
            type: integer
        - name: mes
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LancamentoMes'
  '/titulos/lancamentos':
    put:
      tags:
        - Titulos
      summary: Serviço para realizar lançamentos de despesas do mes
      operationId: putLancamentosTitulo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LancamentoMes'
      responses:
        '200':
          description: Sucesso
  '/titulos/lancamentos/{id}':
    delete:
      tags:
        - Titulos
      summary: Serviço para realizar lançamentos de despesas do mes
      operationId: deleteLancamentosTitulo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Sucesso
  '/titulos/lancamentos/pagar':
    put:
      tags:
        - Titulos
      summary: Serviço para realizar lançamentos de despesas do mes
      operationId: putPagamentoTitulo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PagamentoTitulo'
      responses:
        '200':
          description: Sucesso
components:
  schemas:
    LancamentoMes:
      type: object
      properties:
        id: 
          type: string
        categoriaLancamento:
          type: integer
        classeLancamento:
          type: integer
        nmLancamento:
          type: string
        diaVencimento:
          type: integer
        mesVencimento:
          type: integer
        anoVencimento:
          type: integer
        dataPagamento:
          type: string
          format: date
        valorPlanejado:
          type: number
        valorEfetivo:
          type: number
        statusLancamento:
          type: string
          enum: 
            - PENDENTE
            - PAGO
            - ATRASADO
    PagamentoTitulo:
      type: object
      properties:
        id: 
          type: string
        dataPagamento:
          type: string
          format: date
        valorPago:
          type: number
        valorTitulo:
          type: number